fix: add members on teams create
